hal: infineon: Add option to disable SDIO pullups during SPI sleep
Currently, the SDIO pullups are always set during SPI sleep.
This is a board specific feature which is not always required and
can cause issues when interacting with other pull ups and pull
downs on the line. This commit adds a macro to disable this feature.
Enabling is left as the default so as to prevent a breaking change.

Signed-off-by: Magpie Embedded <magpieembedded@gmail.com>